remove paddle.fluid.layers.layer_norm
