+++ python/paddle/fluid/layers/nn.py
@@ -66,7 +66,6 @@ __all__ = [
     'fc',
     'embedding',
     'row_conv',
-    'layer_norm',
     'spectral_norm',
     'one_hot',
     'autoincreased_step_counter',
@@ -742,142 +741,6 @@ def _pull_box_sparse(
     if len(outs) == 1:
         return outs[0]
     return outs
-
-
-@templatedoc()
-def layer_norm(
-    input,
-    scale=True,
-    shift=True,
-    begin_norm_axis=1,
-    epsilon=1e-05,
-    param_attr=None,
-    bias_attr=None,
-    act=None,
-    name=None,
-):
-    r"""
-    :api_attr: Static Graph
-
-    **Layer Normalization Layer**
-
-    The API implements the function of the Layer Normalization Layer and can be applied to mini-batch input data.
-    Refer to `Layer Normalization <https://arxiv.org/pdf/1607.06450v1.pdf>`_
-
-    The formula is as follows:
-
-    ..  math::
-
-        \\mu & = \\frac{1}{H}\\sum_{i=1}^{H} x_i
-
-        \\sigma & = \\sqrt{\\frac{1}{H}\sum_{i=1}^{H}{(x_i - \\mu)^2} + \\epsilon}
-
-        y & = f(\\frac{g}{\\sigma}(x - \\mu) + b)
-
-    - :math:`x`: the vector representation of the summed inputs to the neurons in that layer.
-    - :math:`H`: the number of hidden units in a layers
-    - :math:`\\epsilon`: the small value added to the variance to prevent division by zero.
-    - :math:`g`: the trainable scale parameter.
-    - :math:`b`: the trainable bias parameter.
-
-    Args:
-        input(Tensor): A multi-dimension ``Tensor`` , and the data type is float32 or float64.
-        scale(bool, optional): Whether to learn the adaptive gain :math:`g` after
-            normalization. Default: True.
-        shift(bool, optional): Whether to learn the adaptive bias :math:`b` after
-            normalization. Default: True.
-        begin_norm_axis(int, optional): The normalization will be performed along
-            dimensions from :attr:`begin_norm_axis` to :attr:`rank(input)`.
-            Default: 1.
-        epsilon(float, optional): The small value added to the variance to prevent
-            division by zero. Default: 1e-05.
-        param_attr(ParamAttr, optional): The parameter attribute for the learnable
-            gain :math:`g`. If :attr:`scale` is False, :attr:`param_attr` is
-            omitted. If :attr:`scale` is True and :attr:`param_attr` is None,
-            a default :code:`ParamAttr` would be added as scale. The
-            :attr:`param_attr` is initialized as 1 if it is added. Default: None.
-        bias_attr(ParamAttr, optional): The parameter attribute for the learnable
-            bias :math:`b`. If :attr:`shift` is False, :attr:`bias_attr` is
-            omitted. If :attr:`shift` is True and :attr:`param_attr` is None,
-            a default :code:`ParamAttr` would be added as bias. The
-            :attr:`bias_attr` is initialized as 0 if it is added. Default: None.
-        act(str, optional): Activation to be applied to the output of layer normalization.
-                  Default: None.
-        name(str): The default value is None.  Normally there is no need for user to set this property.  For more information, please refer to :ref:`api_guide_Name` .
-
-    Returns:
-        Tensor: ``Tensor``  indicating the normalized result, the data type is the same as  ``input`` , and the return dimension is the same as  ``input`` .
-
-    Examples:
-
-        .. code-block:: python
-
-            import paddle
-            paddle.enable_static()
-            x = paddle.static.data(name='x', shape=[8, 32, 32], dtype='float32')
-            output = paddle.static.nn.layer_norm(input=x, begin_norm_axis=1)
-            print(output.shape)  # [8, 32, 32]
-    """
-    assert (
-        _non_static_mode() is not True
-    ), "please use LayerNorm instead of layer_norm in dygraph mode!"
-    helper = LayerHelper('layer_norm', **locals())
-    check_variable_and_dtype(
-        input, 'input', ['float32', 'float64'], 'layer_norm'
-    )
-    dtype = helper.input_dtype()
-
-    # create intput and parameters
-    inputs = {'X': input}
-    input_shape = input.shape
-    param_shape = [reduce(lambda x, y: x * y, input_shape[begin_norm_axis:])]
-    if scale:
-        assert (
-            param_attr is not False
-        ), "param_attr should not be False when using scale."
-        scale = helper.create_parameter(
-            attr=helper.param_attr,
-            shape=param_shape,
-            dtype=dtype,
-            default_initializer=Constant(1.0),
-        )
-        inputs['Scale'] = scale
-    else:
-        if param_attr:
-            warnings.warn("param_attr is only available with scale is True.")
-    if shift:
-        assert (
-            bias_attr is not False
-        ), "bias_attr should not be False when using shift."
-        bias = helper.create_parameter(
-            attr=helper.bias_attr, shape=param_shape, dtype=dtype, is_bias=True
-        )
-        inputs['Bias'] = bias
-    else:
-        if bias_attr:
-            warnings.warn("bias_attr is only available with shift is True.")
-
-    # create output
-    mean_out = helper.create_variable_for_type_inference(
-        dtype=dtype, stop_gradient=True
-    )
-    variance_out = helper.create_variable_for_type_inference(
-        dtype=dtype, stop_gradient=True
-    )
-    layer_norm_out = helper.create_variable_for_type_inference(dtype)
-
-    helper.append_op(
-        type="layer_norm",
-        inputs=inputs,
-        outputs={
-            "Y": layer_norm_out,
-            "Mean": mean_out,
-            "Variance": variance_out,
-        },
-        attrs={"epsilon": epsilon, "begin_norm_axis": begin_norm_axis},
-    )
-
-    return helper.append_activation(layer_norm_out)
 
 
 @templatedoc()
